No to "pause [0|1]" toggles pause
old: pause 0|1
new: pause [0|1]

It is simpler to have a toggle for these so they can be implemented as a
single button (as is very often done). On my keyboard, Fn+f5 is that
button.

pause <no argument>

This change is therefore backward compatible.

Doing this outside of sad required some kind of locking mechanism to
keep track, and is prone to timing-related errors or failing to remove
the lock.